# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Build modern Go apps",
        "title": "Fire-Go",
        "contact": {}
    },
    "host": "https://cprime50.github.io/FireGo-swagger/",
    "basePath": "/api",
    "paths": {
        "/admin/make": {
            "post": {
                "description": "This endpoint allows an admin to promote a user to an admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Make user admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The email of the user to be promoted to admin.",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.EmailInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\\\"message\\\":\\\"User example@example.com is now an admin\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "{\\\"error\\\":\\\"error validating email: invalid format\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\\\"error\\\":\\\"error making admin:\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/remove": {
            "delete": {
                "description": "This endpoint allows an admin to revoke admin rights from a user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Remove user admin rights",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The email of the user to have admin rights revoked.",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.EmailInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\\\"message\\\":\\\"User example@example.com admin rights have been revoked\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "{\\\"error\\\":\\\"error validating email: invalid format\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\\\"error\\\":\\\"error removing admin rights:\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.EmailInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
    }
}